Added ValidateSchema{keyspace, shard} to grpcvtctldserver.
