+++ go/vt/wrangler/schema.go
@@ -20,7 +20,6 @@ import (
 	"bytes"
 	"fmt"
 	"html/template"
-	"sort"
 	"sync"
 	"time"
 
@@ -33,6 +32,7 @@ import (
 	"vitess.io/vitess/go/vt/mysqlctl/tmutils"
 	"vitess.io/vitess/go/vt/topo"
 	"vitess.io/vitess/go/vt/topo/topoproto"
+	"vitess.io/vitess/go/vt/vtctl/grpcvtctldserver"
 
 	tabletmanagerdatapb "vitess.io/vitess/go/vt/proto/tabletmanagerdata"
 	topodatapb "vitess.io/vitess/go/vt/proto/topodata"
@@ -144,176 +144,18 @@ func (wr *Wrangler) diffSchema(ctx context.Context, primarySchema *tabletmanager
 
 // ValidateSchemaShard will diff the schema from all the tablets in the shard.
 func (wr *Wrangler) ValidateSchemaShard(ctx context.Context, keyspace, shard string, excludeTables []string, includeViews bool, includeVSchema bool) error {
-	si, err := wr.ts.GetShard(ctx, keyspace, shard)
-	if err != nil {
-		return fmt.Errorf("GetShard(%v, %v) failed: %v", keyspace, shard, err)
-	}
-
-	// get schema from the primary, or error
-	if !si.HasPrimary() {
-		return fmt.Errorf("no primary in shard %v/%v", keyspace, shard)
-	}
-	log.Infof("Gathering schema for primary %v", topoproto.TabletAliasString(si.PrimaryAlias))
-	primarySchema, err := wr.GetSchema(ctx, si.PrimaryAlias, nil, excludeTables, includeViews)
-	if err != nil {
-		return fmt.Errorf("GetSchema(%v, nil, %v, %v) failed: %v", si.PrimaryAlias, excludeTables, includeViews, err)
-	}
-
-	if includeVSchema {
-		err := wr.ValidateVSchema(ctx, keyspace, []string{shard}, excludeTables, includeViews)
-		if err != nil {
-			return err
-		}
-	}
-
-	// read all the aliases in the shard, that is all tablets that are
-	// replicating from the primary
-	aliases, err := wr.ts.FindAllTabletAliasesInShard(ctx, keyspace, shard)
-	if err != nil {
-		return fmt.Errorf("FindAllTabletAliasesInShard(%v, %v) failed: %v", keyspace, shard, err)
-	}
-
-	// then diff with all replicas
-	er := concurrency.AllErrorRecorder{}
-	wg := sync.WaitGroup{}
-	for _, alias := range aliases {
-		if topoproto.TabletAliasEqual(alias, si.PrimaryAlias) {
-			continue
-		}
-
-		wg.Add(1)
-		go wr.diffSchema(ctx, primarySchema, si.PrimaryAlias, alias, excludeTables, includeViews, &wg, &er)
-	}
-	wg.Wait()
-	if er.HasErrors() {
-		return fmt.Errorf("schema diffs: %v", er.Error().Error())
-	}
-	return nil
+	return grpcvtctldserver.NewVtctldServer(wr.ts).ValidateSchemaShard(ctx, keyspace, shard, excludeTables, includeViews, includeVSchema)
 }
 
 // ValidateSchemaKeyspace will diff the schema from all the tablets in
 // the keyspace.
 func (wr *Wrangler) ValidateSchemaKeyspace(ctx context.Context, keyspace string, excludeTables []string, includeViews, skipNoPrimary bool, includeVSchema bool) error {
-	// find all the shards
-	shards, err := wr.ts.GetShardNames(ctx, keyspace)
-	if err != nil {
-		return fmt.Errorf("GetShardNames(%v) failed: %v", keyspace, err)
-	}
-
-	// corner cases
-	if len(shards) == 0 {
-		return fmt.Errorf("no shards in keyspace %v", keyspace)
-	}
-	sort.Strings(shards)
-	if len(shards) == 1 {
-		return wr.ValidateSchemaShard(ctx, keyspace, shards[0], excludeTables, includeViews, includeVSchema)
-	}
-
-	var referenceSchema *tabletmanagerdatapb.SchemaDefinition
-	var referenceAlias *topodatapb.TabletAlias
-
-	// then diff with all other tablets everywhere
-	er := concurrency.AllErrorRecorder{}
-	wg := sync.WaitGroup{}
-
-	// If we are checking against the vschema then all shards
-	// should just be validated individually against it
-	if includeVSchema {
-		err := wr.ValidateVSchema(ctx, keyspace, shards, excludeTables, includeViews)
-		if err != nil {
-			return err
-		}
-	}
-
-	// then diffs all tablets in the other shards
-	for _, shard := range shards[0:] {
-		si, err := wr.ts.GetShard(ctx, keyspace, shard)
-		if err != nil {
-			er.RecordError(fmt.Errorf("GetShard(%v, %v) failed: %v", keyspace, shard, err))
-			continue
-		}
-
-		if !si.HasPrimary() {
-			if !skipNoPrimary {
-				er.RecordError(fmt.Errorf("no primary in shard %v/%v", keyspace, shard))
-			}
-			continue
-		}
-
-		if referenceSchema == nil {
-			referenceAlias = si.PrimaryAlias
-			log.Infof("Gathering schema for reference primary %v", topoproto.TabletAliasString(referenceAlias))
-			referenceSchema, err = wr.GetSchema(ctx, referenceAlias, nil, excludeTables, includeViews)
-			if err != nil {
-				return fmt.Errorf("GetSchema(%v, nil, %v, %v) failed: %v", referenceAlias, excludeTables, includeViews, err)
-			}
-		}
-
-		aliases, err := wr.ts.FindAllTabletAliasesInShard(ctx, keyspace, shard)
-		if err != nil {
-			er.RecordError(fmt.Errorf("FindAllTabletAliasesInShard(%v, %v) failed: %v", keyspace, shard, err))
-			continue
-		}
-
-		for _, alias := range aliases {
-			// Don't diff schemas for self
-			if referenceAlias == alias {
-				continue
-			}
-			wg.Add(1)
-			go wr.diffSchema(ctx, referenceSchema, referenceAlias, alias, excludeTables, includeViews, &wg, &er)
-		}
-	}
-	wg.Wait()
-	if er.HasErrors() {
-		return fmt.Errorf("schema diffs: %v", er.Error().Error())
-	}
-	return nil
+	return grpcvtctldserver.NewVtctldServer(wr.ts).ValidateSchemaKeyspace(ctx, keyspace, excludeTables, includeViews, skipNoPrimary, includeVSchema)
 }
 
 // ValidateVSchema compares the schema of each primary tablet in "keyspace/shards..." to the vschema and errs if there are differences
 func (wr *Wrangler) ValidateVSchema(ctx context.Context, keyspace string, shards []string, excludeTables []string, includeViews bool) error {
-	vschm, err := wr.ts.GetVSchema(ctx, keyspace)
-	if err != nil {
-		return fmt.Errorf("GetVSchema(%s) failed: %v", keyspace, err)
-	}
-
-	shardFailures := concurrency.AllErrorRecorder{}
-	var wg sync.WaitGroup
-	wg.Add(len(shards))
-
-	for _, shard := range shards {
-		go func(shard string) {
-			defer wg.Done()
-			notFoundTables := []string{}
-			si, err := wr.ts.GetShard(ctx, keyspace, shard)
-			if err != nil {
-				shardFailures.RecordError(fmt.Errorf("GetShard(%v, %v) failed: %v", keyspace, shard, err))
-				return
-			}
-			primarySchema, err := wr.GetSchema(ctx, si.PrimaryAlias, nil, excludeTables, includeViews)
-			if err != nil {
-				shardFailures.RecordError(fmt.Errorf("GetSchema(%s, nil, %v, %v) (%v/%v) failed: %v", si.PrimaryAlias.String(),
-					excludeTables, includeViews, keyspace, shard, err,
-				))
-				return
-			}
-			for _, tableDef := range primarySchema.TableDefinitions {
-				if _, ok := vschm.Tables[tableDef.Name]; !ok {
-					notFoundTables = append(notFoundTables, tableDef.Name)
-				}
-			}
-			if len(notFoundTables) > 0 {
-				shardFailure := fmt.Errorf("%v/%v has tables that are not in the vschema: %v", keyspace, shard, notFoundTables)
-				shardFailures.RecordError(shardFailure)
-			}
-		}(shard)
-	}
-	wg.Wait()
-	if shardFailures.HasErrors() {
-		return fmt.Errorf("ValidateVSchema(%v, %v, %v, %v) failed: %v", keyspace, shards, excludeTables, includeViews, shardFailures.Error().Error())
-	}
-	return nil
+	return grpcvtctldserver.NewVtctldServer(wr.ts).ValidateVSchema(ctx, keyspace, shards, excludeTables, includeViews)
 }
 
 // PreflightSchema will try a schema change on the remote tablet.
